requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
MouseInfo==0.1.3
openai==1.30.1
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
PyAutoGUI==0.9.54
pydantic==2.7.1
pydantic_core==2.18.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Quartz==10.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dotenv==1.0.1
pytweening==1.2.0
PyVirtualDisplay==3.0
requests==2.31.0
rubicon-objc==0.4.9
selenium==4.20.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
